+++ tests/dialects/test_snowflake.py
@@ -143,6 +143,19 @@ class TestSnowflake(Validator):
                 "snowflake": r"SELECT 'a \' \\ \\t \\x21 z $ '",
             },
         )
+        self.validate_identity("SELECT REGEXP_LIKE(a, b, c)")
+        self.validate_all(
+            "SELECT RLIKE(a, b)",
+            write={
+                "snowflake": "SELECT REGEXP_LIKE(a, b)",
+            },
+        )
+        self.validate_all(
+            "SELECT a FROM test SAMPLE BLOCK (0.5) SEED (42)",
+            write={
+                "snowflake": "SELECT a FROM test TABLESAMPLE BLOCK (0.5) SEED (42)",
+            },
+        )
 
     def test_null_treatment(self):
         self.validate_all(
@@ -218,5 +231,26 @@ class TestSnowflake(Validator):
             "SELECT DATE_PART(month FROM a::DATETIME)",
             write={
                 "snowflake": "SELECT EXTRACT(month FROM CAST(a AS DATETIME))",
+            },
+        )
+
+    def test_semi_structured_types(self):
+        self.validate_identity("SELECT CAST(a AS VARIANT)")
+        self.validate_all(
+            "SELECT a::VARIANT",
+            write={
+                "snowflake": "SELECT CAST(a AS VARIANT)",
+                "tsql": "SELECT CAST(a AS SQL_VARIANT)",
+            },
+        )
+        self.validate_identity("SELECT CAST(a AS ARRAY)")
+        self.validate_all(
+            "ARRAY_CONSTRUCT(0, 1, 2)",
+            write={
+                "snowflake": "[0, 1, 2]",
+                "bigquery": "[0, 1, 2]",
+                "duckdb": "LIST_VALUE(0, 1, 2)",
+                "presto": "ARRAY[0, 1, 2]",
+                "spark": "ARRAY(0, 1, 2)",
             },
         )