+++ tests/test_expressions.py
@@ -224,9 +224,6 @@ class TestExpressions(unittest.TestCase):
         self.assertEqual(actual_expression_2.sql(dialect="presto"), "IF(c - 2 > 0, c - 2, b)")
         self.assertIs(actual_expression_2, expression)
 
-        with self.assertRaises(ValueError):
-            parse_one("a").transform(lambda n: None)
-
     def test_transform_no_infinite_recursion(self):
         expression = parse_one("a")
 
@@ -246,6 +243,35 @@ class TestExpressions(unittest.TestCase):
             return node
 
         self.assertEqual(expression.transform(fun).sql(), "SELECT a, b FROM x")
+
+    def test_transform_node_removal(self):
+        expression = parse_one("SELECT a, b FROM x")
+
+        def remove_column_b(node):
+            if isinstance(node, exp.Column) and node.name == "b":
+                return None
+            return node
+
+        self.assertEqual(expression.transform(remove_column_b).sql(), "SELECT a FROM x")
+        self.assertEqual(expression.transform(lambda _: None), None)
+
+        expression = parse_one("CAST(x AS FLOAT)")
+
+        def remove_non_list_arg(node):
+            if isinstance(node, exp.DataType):
+                return None
+            return node
+
+        self.assertEqual(expression.transform(remove_non_list_arg).sql(), "CAST(x AS )")
+
+        expression = parse_one("SELECT a, b FROM x")
+
+        def remove_all_columns(node):
+            if isinstance(node, exp.Column):
+                return None
+            return node
+
+        self.assertEqual(expression.transform(remove_all_columns).sql(), "SELECT FROM x")
 
     def test_replace(self):
         expression = parse_one("SELECT a, b FROM x")
@@ -290,6 +316,7 @@ class TestExpressions(unittest.TestCase):
         self.assertIsInstance(parse_one("MAX(a)"), exp.Max)
         self.assertIsInstance(parse_one("MIN(a)"), exp.Min)
         self.assertIsInstance(parse_one("MONTH(a)"), exp.Month)
+        self.assertIsInstance(parse_one("POSITION(' ' IN a)"), exp.StrPosition)
         self.assertIsInstance(parse_one("POW(a, 2)"), exp.Pow)
         self.assertIsInstance(parse_one("POWER(a, 2)"), exp.Pow)
         self.assertIsInstance(parse_one("QUANTILE(a, 0.90)"), exp.Quantile)
@@ -388,3 +415,34 @@ class TestExpressions(unittest.TestCase):
         self.assertEqual(parse_one("heLLO()").sql(normalize_functions=None), "heLLO()")
         self.assertEqual(parse_one("SUM(x)").sql(normalize_functions="lower"), "sum(x)")
         self.assertEqual(parse_one("sum(x)").sql(normalize_functions="upper"), "SUM(x)")
+
+    def test_properties_from_dict(self):
+        self.assertEqual(
+            exp.Properties.from_dict(
+                {
+                    "FORMAT": "parquet",
+                    "PARTITIONED_BY": [exp.to_identifier("a"), exp.to_identifier("b")],
+                    "custom": 1,
+                    "TABLE_FORMAT": exp.to_identifier("test_format"),
+                    "ENGINE": None,
+                    "COLLATE": True,
+                }
+            ),
+            exp.Properties(
+                expressions=[
+                    exp.FileFormatProperty(this=exp.Literal.string("FORMAT"), value=exp.Literal.string("parquet")),
+                    exp.PartitionedByProperty(
+                        this=exp.Literal.string("PARTITIONED_BY"),
+                        value=exp.Tuple(expressions=[exp.to_identifier("a"), exp.to_identifier("b")]),
+                    ),
+                    exp.AnonymousProperty(this=exp.Literal.string("custom"), value=exp.Literal.number(1)),
+                    exp.TableFormatProperty(
+                        this=exp.Literal.string("TABLE_FORMAT"), value=exp.to_identifier("test_format")
+                    ),
+                    exp.EngineProperty(this=exp.Literal.string("ENGINE"), value=exp.NULL),
+                    exp.CollateProperty(this=exp.Literal.string("COLLATE"), value=exp.TRUE),
+                ]
+            ),
+        )
+
+        self.assertRaises(ValueError, exp.Properties.from_dict, {"FORMAT": {"key": "value"}})