+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+v6.2.0
+------
+
+Changes:
+
+- New: TSQL support
+
+- Breaking: Removed $ from tokenizer, added @ placeholders
+
+- Improvement: Nodes can now be removed in transform and replace [8cd81c3](https://github.com/tobymao/sqlglot/commit/8cd81c36561463b9849a8e0c2d70248c5b1feb62)
+
+- Improvement: Snowflake timestamp support
+
+- Improvement: Property conversion for CTAS Builder
+
+- Improvement: Tokenizers are now unique per dialect instance
+
 v6.1.0
 ------